moved code into cppformat subdirectory
Proxy headers have been placed into the project root to emit
deprecation warnings.
