+++ test/format-impl-test.cc
@@ -29,7 +29,7 @@
 #include "test-assert.h"
 
 // Include format.cc instead of format.h to test implementation-specific stuff.
-#include "format.cc"
+#include <cppformat/format.cc>
 
 #include <cstring>
